% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atp.MVCN.R
\name{atp.MVCN}
\alias{atp.MVCN}
\title{Atypical Detection Points Using Matrix-Variate Contaminated Normal Hidden Markov Models}
\usage{
atp.MVCN(Y, pgood, class)
}
\arguments{
\item{Y}{An array with dimensions \code{p} x \code{r} x \code{num} x \code{t}, where \code{p} is the number of
variables in the rows of each data matrix, \code{r} is the number of variables in the columns of each
data matrix, \code{num} is the number of data observations, and \code{t} is the number of time points.}

\item{pgood}{An array with dimensions \code{num} x \code{t} x \code{k} containing the estimated probability of being typical for each point, given the time and state.}

\item{class}{An \code{num} x \code{t} matrix containing the state memberships.}
}
\value{
An \code{num} x \code{t} matrix containing, for each observation and time, a 0 if it that matrix is typical and 1 otherwise.
}
\description{
Detects atypical matrices via matrix-variate contaminated normal Hidden Markov Models.
}
\examples{
data("simData2")
Y <- simData2$Y
init <- Eigen.HMM_init(Y = Y, k = 2, density = "MVCN", mod.row = "EEE", mod.col = "EE", nstartR = 1)
fit <- Eigen.HMM_fit(Y = Y, init.par = init, nThreads = 1)
atp <- atp.MVCN(Y = Y,
               pgood = fit[["results"]][[1]][[1]][[1]][["pgood"]],
               class = fit[["results"]][[1]][[1]][[1]][["class"]])
which(atp==1)
which(simData2[["atp.tr"]]==1)
}
