% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simData.R
\docType{data}
\name{simData}
\alias{simData}
\title{A Simulated Dataset from a Matrix-Variate t Hidden Markov Model}
\format{
A list containing two elements:
\describe{
\item{1)}{An array with \code{p = 2} variables in the rows, \code{r = 3} variables in the columns, \code{num = 50} matrices, and \code{t = 3} time points.}
\item{2)}{An \code{num} x \code{t} matrix containing the state memberships.}
}
}
\usage{
data(simData)
}
\description{
A simulated dataset generated from a matrix-variate t Hidden Markov Model with 2 states and an EE - EE covariance structure.
}
\keyword{datasets}
